+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Input: add `cleanUpOnChange`, `cleanUpOnValidate` props.
+
 ## 0.18.3 (2019-8-13)
 
 ## 0.18.2 (2019-8-13)